# expression: add explicit timer for sleep evaluate