<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
Currently `sleep` expression is evaluation by a time ticker with duration of 10ms, but in some case, the sleep time is much smaller (mostly in unit test like [here](https://github.com/pingcap/tidb/blob/master/expression/builtin_miscellaneous_vec_test.go#L37)) than this tick duration, we will still need to sleep for 10ms.

### What is changed and how it works?
~~Make sleep duration same as the sleep time if it is smaller than 10ms, thus the sleep expression can be more accurate and it will make some test run mush faster.~~
Updated for @SunRunAway review comment, use the origin ticker for session killed check, and a new timer for the sleep check, we can gain better accuracy and performance. 

Before:
```
$ go test github.com/pingcap/tidb/expression
PASS: builtin_miscellaneous_vec_test.go:104: testEvaluatorSuite.TestVectorizedBuiltinMiscellaneousEvalOneVec	22.056s
PASS: builtin_miscellaneous_vec_test.go:108: testEvaluatorSuite.TestVectorizedBuiltinMiscellaneousFunc	22.933s
ok  	github.com/pingcap/tidb/expression	52.283s
```
After:
```
$ go test github.com/pingcap/tidb/expression
PASS: builtin_miscellaneous_vec_test.go:104: testEvaluatorSuite.TestVectorizedBuiltinMiscellaneousEvalOneVec	0.665s
PASS: builtin_miscellaneous_vec_test.go:108: testEvaluatorSuite.TestVectorizedBuiltinMiscellaneousFunc	0.678s
ok  	github.com/pingcap/tidb/expression	13.019s
```

### Related changes

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test
- Integration test

Side effects

### Release note <!-- bugfixes or new feature need a release note -->